+++ store/store.ts
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { completedTasksReducer } from './completedTasks/completedTasksSlice';
-import { householdReducer } from './households/householdsSlice';
-import { memberReducer } from './members/membersSlice';
-import { requestReducer } from './request/requestsSlice';
+import { householdReducer as householdsReducer } from './households/householdsSlice';
+import { memberReducer as membersReducer } from './members/membersSlice';
+import { requestReducer as requestsReducer } from './requests/requestsSlice';
 import { scheduledTasksReducer } from './scheduledTasks/scheduledTasksSlice';
 import { tasksReducer } from './tasks/tasksSlice';
 import userReducer from './user/userReducer';
@@ -11,11 +11,11 @@ export const store = configureStore({
   reducer: {
     user: userReducer,
     tasks: tasksReducer,
-    members: memberReducer,
-    households: householdReducer,
+    members: membersReducer,
+    households: householdsReducer,
     completedTasks: completedTasksReducer,
     scheduledTasks: scheduledTasksReducer,
-    requests: requestReducer,
+    requests: requestsReducer,
   },
 });
